refactor: simplify withdrawal amount to Long and verify HTTP responses in concurrency test

+++ src/main/kotlin/com/wannabe/wallet/application/wallet/assembler/WalletWithdrawalAssembler.kt
@@ -1,12 +1,11 @@
 package com.wannabe.wallet.application.wallet.assembler
 
 import com.wannabe.wallet.application.wallet.dto.WithdrawalResultDTO
-import java.math.BigDecimal
 
 interface WalletWithdrawalAssembler {
     fun withdraw(
         walletId: String,
-        amount: BigDecimal,
+        amount: Long,
         currency: String,
         transactionId: String,
     ): WithdrawalResultDTO
